Transform feedback to be continued in dev-0.8

+++ src/Evaluation.h
@@ -82,8 +82,8 @@ struct EvaluationInfo
 	float pad2[4];
 	
 	float viewport[2];
-	size_t mFrame;
-	size_t mLocalFrame;
+	int mFrame;
+	int mLocalFrame;
 };
 
 struct TextureFormat
